moved list color to outside of list. set default height.

+++ app/assets/javascripts/shop.js
@@ -45,34 +45,20 @@ var shop = {
   },
 
   addItem: function() {
-    var listId = $("#lists").val();
-
-    // $.ajax({
-    //   url: "/lists/" + listId + "/items.json",
-    //   dataType: "json",
-    //   type: "POST",
-    //   data: { item : { list_id : listId,  name : $("#addItemName").val() } },
-    //   success: function(response) {
-    //     $("#emptyItem").remove();
-    //     // Create new item.
-    //     $("#itemList")
-    //       .append($("<li id='item-" + response.id + "'>")
-    //         .append($("<span class='itemName'>")
-    //           .append(response.name))
-    //         .append("<button class='deleteItem'>X</button>"));
-    //   }
-    // });
-
-    $.create("/lists/" + listId + "/items/", { item : { list_id : listId,  name : $("#addItemName").val() } },
-        function(response) {
-      $("#emptyItem").remove();
-      // Create new item.
-      $("#itemList")
-        .append($("<li id='item-" + response.id + "'>")
-          .append($("<span class='itemName'>")
-            .append(response.name))
-          .append("<button class='deleteItem'>X</button>"));
-    }, "json");
+    var newItemName = $.trim($("#addItemName").val());
+    if (newItemName != "") {
+      var listId = $("#lists").val();
+      $.create("/lists/" + listId + "/items/", { item : { list_id : listId,  name : $("#addItemName").val() } },
+          function(response) {
+        $("#emptyItem").remove();
+        // Create new item.
+        $("#itemList")
+          .append($("<li id='item-" + response.id + "'>")
+            .append($("<span class='itemName'>")
+              .append(response.name))
+            .append("<button class='deleteItem'>X</button>"));
+      }, "json");
+    }
     $("#addItemName").val("");
   }
 };
